requirements.txt
# ==========================================================
# TENDER INTELLIGENCE SYSTEM - DEPENDENCIES
# ==========================================================

# Web Framework
flask==3.0.0
gunicorn==21.2.0

# Scraping
requests==2.31.0
beautifulsoup4==4.12.2
lxml==4.9.3

# Selenium (for National Treasury - optional on cloud)
selenium==4.15.2

# Excel handling
openpyxl==3.1.2

# Configuration
PyYAML==6.0.1

# Utilities
python-dateutil==2.8.2
urllib3==2.1.0
